Towers Watson coding test
I started using Bootstrap and KnockoutJS to help me, so I would focus on
design and backend development.  That file is the reg.  I made one that
only used jQuery and KnockoutJS and pulled the classes from Bootstrap
that I only used another file.  I wasn't sure if I was allowed to use
web technologies at all for the coding test or be able to access them.